Have Coin Pick Up and Shopping list done
just need to add all the shopping to the list that we want in the game and then tweek the code to fit. when the play hits the bread pick up the player will get two score and the bread will cross off on the list, when the player hits the mix veg both the carrots and the peas will be crossed off and the player will only get a score of one.